Fixed some little issues.
Forced ydl to downdload the playlist if the ID refers both to a playlist and to a video.
Fixed inconsistency with README.md about playlist default directory path: now when no directory is provided for a playlist, the songs are downloaded in the ./PLAYLIST_ID/ subfolder.
Now also removing .ytdl files when the program stops because of an exception.